// src/apis/models/MatchHalfTimeRequestHalfType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * MatchDay Swagger
 * MatchDay 요구사항에 대한 REST API
 * OpenAPI spec version: 0.0.1
 */

/**
 * 전/후반
 */
export type MatchHalfTimeRequestHalfType =
  (typeof MatchHalfTimeRequestHalfType)[keyof typeof MatchHalfTimeRequestHalfType];

export const MatchHalfTimeRequestHalfType = {
  FIRST_HALF: 'FIRST_HALF',
  SECOND_HALF: 'SECOND_HALF',
} as const;