// src/apis/models/MatchHalfTimeRequestTimeType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * MatchDay Swagger
 * MatchDay 요구사항에 대한 REST API
 * OpenAPI spec version: 0.0.1
 */

/**
 * 시작/종료 시간
 */
export type MatchHalfTimeRequestTimeType =
  (typeof MatchHalfTimeRequestTimeType)[keyof typeof MatchHalfTimeRequestTimeType];

export const MatchHalfTimeRequestTimeType = {
  START_TIME: 'START_TIME',
  END_TIME: 'END_TIME',
} as const;